pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "py-context-graph"
version = "0.1.0"
description = "Extract, enrich, cluster, and query decisions from unstructured conversations using LLMs."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "ResearchifyLabs" },
]
keywords = ["decisions", "llm", "knowledge-graph", "clustering", "context", "conversations"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pydantic>=2.0",
]

[project.optional-dependencies]
llm = ["litellm"]
firestore = ["google-cloud-firestore"]
memory = ["pandas"]
all = ["litellm", "google-cloud-firestore", "pandas"]
dev = [
    "litellm",
    "pandas",
    "google-cloud-firestore",
]

[project.urls]
Homepage = "https://github.com/ResearchifyLabs/py-context-graph"
Repository = "https://github.com/ResearchifyLabs/py-context-graph"
Issues = "https://github.com/ResearchifyLabs/py-context-graph/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/decision_graph"]
